Controller

+++ app/Http/Controllers/V1/Cecy/ParticipantController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\V1\Cecy\Participants\DestroysParticipantRequest;
 use App\Http\Requests\V1\Cecy\Participants\StoreParticipantRequest;
 use App\Http\Requests\V1\Cecy\Planifications\IndexPlanificationRequest;
+use App\Http\Requests\V1\Cecy\Planifications\DestroysPlanificationRequest;
 use App\Http\Requests\V1\Cecy\Registrations\RegistrationStateModificationRequest;
 use App\Http\Requests\V1\Cecy\Participants\StoreParticipantUserRequest;
 use App\Http\Resources\V1\Cecy\Participants\ParticipantResource;
@@ -194,8 +195,52 @@ class ParticipantController extends Controller
         return 'por revisar';
     }
 
-    public function destroyParticipant(DestroyParticipantRequest $request, Participant $participant)
+    //Modificacion de Participante
+    public function chageDataParticipante(/*ChageDataParticipantRequest*/ $request, Participant $participant){
+        $participant->identificationType()->associate(Catalogue::find($request->input('identificationType.id')));
+        $participant->sex()->associate(Catalogue::find($request->input('sex.id')));
+        $participant->gender()->associate(Catalogue::find($request->input('gender.id')));
+        $participant->bloodType()->associate(Catalogue::find($request->input('bloodType.id')));
+        $participant->ethnicOrigin()->associate(Catalogue::find($request->input('ethnicOrigin.id')));
+        $participant->civilStatus()->associate(Catalogue::find($request->input('civilStatus.id')));
+
+        $participant->username = $request->input('username');
+        $participant->name = $request->input('name');
+        $participant->lastname = $request->input('lastname');
+        $participant->birthdate = $request->input('birthdate');
+        $participant->email = $request->input('email');
+
+        $participant->save();
+        $participant->addEmails($request->input('emails'));
+
+        return (new UserResource($participant))
+            ->additional([
+                'msg' => [
+                    'summary' => 'Usuario Actualizado',
+                    'detail' => '',
+                    'code' => '200'
+                ]
+            ])
+            ->response()->setStatusCode(201);
+    }
+    //Aceptacion de Participante
+    public function acceptParticipante(/*AcceptParticipantRequest*/ $request, Participant $participant){
+
+    }
+
+    //Eliminacion de Participante
+    public function destroyParticipant(/*DestroyParticipantRequest*/ $request, Participant $participant)
     {
+        if ($request->user()->id === $participant->id) {
+            return response()->json([
+                'msg' => [
+                    'summary' => 'Error al eliminar',
+                    'detail' => 'El usuario se encuentra logueado',
+                    'code' => '400'
+                ],
+            ], 400);
+        }
+
         $participant->delete();
 
         return (new ParticipantResource($participant))
